+++ CHANGELOG.md
@@ -29,6 +29,7 @@ scheme are considered to be bugs.
 ### Fixed
 
 * [#402](https://github.com/intridea/hashie/pull/402): Use a Railtie to set Hashie.logger on rails boot - [@matthewrudy](https://github.com/matthewrudy).
+* [#406](https://github.com/intridea/hashie/pull/406): Ensure that subclasses that disable warnings propagate that setting to grandchild classes - [@michaelherold](https://github.com/michaelherold).
 * Your contribution here.
 
 ### Security